Fix problems with go1.7+
There were failures with go1.7 and higher in the TestOnInstanceMethod
test.

First was:

    panic: unknown method no [recovered]
            panic: unknown method no

This appears to be due to https://github.com/golang/go/issues/15673
where unexported methods are made even harder to get to via
reflection.

With that addressed, tests still failed:

    --- FAIL: TestOnInstanceMethod (0.00s)
            Error Trace:    monkey_test.go:94
            Error:          Should be true
            Error Trace:    monkey_test.go:95
            Error:          Should be false

This appears to be due to `reflect.Value` no longer being usable to
uniquely identify a method. Using `.Pointer()` on the `reflect.Value`
gives a workable ID.